requirements.txt
-ip==21.3.1
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
charset-normalizer==2.0.12
discord.py==2.0.0a4197+g14e83dff
environs==9.5.0
frozenlist==1.3.0
idna==3.3
marshmallow==3.15.0
multidict==6.0.2
packaging==21.3
pip==22.0.4
pyparsing==3.0.8
python-dotenv==0.20.0
setuptools==60.2.0
wheel==0.37.1
yarl==1.7.2

vkbottle~=4.3.0